edit items entity and dto to let onHand be float

+++ composeApp/src/commonMain/kotlin/org/abapps/app/data/remote/model/ItemDto.kt
@@ -13,7 +13,7 @@ data class ItemDto(
     val description: String? = null,
     val styleId: Long? = null,
     val styleName: String? = null,
-    val onHand: Int? = null,
+    val onHand: Float? = null,
     val freeCard: Boolean? = null,
     val freeCardPrice: Float? = null,
     val cost: Float? = null,
